refactor: unify document ID retrieval in ObservableSanityClient and SanityClient to dataMethods
